Now automatically building a tower at start of game & detected function causing vector error

+++ BTD6-ai/Action.h
@@ -4,6 +4,6 @@ class Action
 public:
 	bool isReady(double t_currentRound) { return t_currentRound >= round; }
 protected:
-	double round; // Round to do action on
+	int round; // Round to do action on
 };
