+++ BTD6-ai/UpgradeAction.cpp
@@ -1,6 +1,6 @@
 #include "UpgradeAction.h"
 
-UpgradeAction::UpgradeAction(double t_round, Tower* t_tower, short t_path)
+UpgradeAction::UpgradeAction(int t_round, Tower* t_tower, short t_path)
 {
 	round = t_round;
 	tower = t_tower;
